Move handler implementation to mod.rs
