+++ src/handler/standard.rs
@@ -10,7 +10,7 @@ use common::proto;
 use common::proto::organization::Organization_Authorization_Role::TRANSACTOR;
 use state::ConsensourceState;
 
-use transaction_handler::{agent, organization};
+use handler::{agent, organization};
 
 /// Creates a new Standard and submits it to state
 ///
@@ -287,7 +287,7 @@ pub fn make_proto(
 mod tests {
     use super::*;
 
-    use transaction_handler::test_utils::*;
+    use handler::test_utils::*;
 
     #[test]
     /// Test that if CreateStandardAction is valid an OK is returned and a new Standard is added to state
