--- src/transaction_handler/mod.rs
@@ -1,8 +1,0 @@
-pub mod agent;
-pub mod assertion;
-pub mod certificate;
-pub mod factory;
-pub mod handler;
-pub mod organization;
-pub mod standard;
-pub mod test_utils;
